Split notifier in two
