+++ apps/transport/lib/jobs/consolidate_bnlc_job.ex
@@ -163,7 +163,7 @@ defmodule Transport.Jobs.ConsolidateBNLCJob do
 
     file_url = Transport.S3.permanent_url(@s3_bucket, filename)
 
-    email = Transport.UserNotifier.bnlc_consolidation_report(subject, body, file_url)
+    email = Transport.AdminNotifier.bnlc_consolidation_report(subject, body, file_url)
     Transport.Mailer.deliver(email)
   end
