+++ apps/transport/lib/jobs/datasets_climate_resilience_bill_not_lo_licence_job.ex
@@ -13,7 +13,7 @@ defmodule Transport.Jobs.DatasetsClimateResilienceBillNotLOLicenceJob do
     remove_climate_resilience_bill_tag(datasets)
 
     unless Enum.empty?(datasets) do
-      mail = Transport.UserNotifier.datasets_climate_resilience_bill_inappropriate_licence(datasets)
+      mail = Transport.AdminNotifier.datasets_climate_resilience_bill_inappropriate_licence(datasets)
       Transport.Mailer.deliver(mail)
     end
