+++ apps/transport/lib/jobs/datasets_without_gtfs_rt_related_resources_notification_job.ex
@@ -14,7 +14,7 @@ defmodule Transport.Jobs.DatasetsWithoutGTFSRTRelatedResourcesNotificationJob do
   def send_email([]), do: :ok
 
   def send_email(datasets) do
-    email = Transport.UserNotifier.datasets_without_gtfs_rt_related_resources(datasets)
+    email = Transport.AdminNotifier.datasets_without_gtfs_rt_related_resources(datasets)
     Transport.Mailer.deliver(email)
     :ok
   end
