+++ apps/transport/lib/transport/data_checker.ex
@@ -189,7 +189,7 @@ defmodule Transport.DataChecker do
   defp send_outdated_data_mail([] = _records), do: []
 
   defp send_outdated_data_mail(records) do
-    email_content = Transport.UserNotifier.expiration_admin(records)
+    email_content = Transport.AdminNotifier.expiration(records)
     Transport.Mailer.deliver(email_content)
     records
   end
@@ -199,7 +199,9 @@ defmodule Transport.DataChecker do
     do: nil
 
   defp send_inactive_datasets_mail(reactivated_datasets, inactive_datasets, archived_datasets) do
-    email_content = Transport.UserNotifier.inactive_datasets(reactivated_datasets, inactive_datasets, archived_datasets)
+    email_content =
+      Transport.AdminNotifier.inactive_datasets(reactivated_datasets, inactive_datasets, archived_datasets)
+
     Transport.Mailer.deliver(email_content)
   end
 end
