+++ apps/transport/lib/transport_web/live/feedback_live.ex
@@ -72,7 +72,7 @@ defmodule TransportWeb.Live.FeedbackLive do
   @spec deliver_mail(DB.UserFeedback.t()) :: {:ok, term} | {:error, term}
   defp deliver_mail(feedback) do
     feedback_email =
-      Transport.UserNotifier.feedback(feedback.rating, feedback.explanation, feedback.email, feedback.feature)
+      Transport.AdminNotifier.feedback(feedback.rating, feedback.explanation, feedback.email, feedback.feature)
 
     Transport.Mailer.deliver(feedback_email)
   end
